requirements.txt
greenlet==1.0.0
numpy==1.20.1
pandas==1.2.3
psycopg2-binary==2.8.6
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.2
